enable logging out
