+++ pages/api/logout.js
@@ -1,0 +1,6 @@
+import withSession from "../../lib/withSession";
+
+export default withSession(async (req, res) => {
+  req.session.destroy();
+  res.json({ isLoggedIn: false });
+});
